Loading Controller Single Responsibility Principle
