+++ src/main/java/burgerput/project/zenput/repository/foodRepository/CustomFoodRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
+
 public interface CustomFoodRepository extends JpaRepository<CustomFood, Integer> {
     @Modifying(clearAutomatically = true)
     @Query(value = "truncate table Custom_food", nativeQuery = true)
