+++ src/main/java/burgerput/project/zenput/repository/foodRepository/FoodRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
+
 public interface FoodRepository extends JpaRepository<Food,Integer> {
 
     @Modifying(clearAutomatically = true)
