+++ src/main/java/burgerput/project/zenput/repository/machineRepository/CustomMachineRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-@Transactional
+
 public interface CustomMachineRepository extends JpaRepository<CustomMachine,Integer> {
 
     @Modifying(clearAutomatically = true)
