+++ src/main/java/burgerput/project/zenput/repository/machineRepository/MachineRepository.java
@@ -10,7 +10,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-@Repository
 public interface MachineRepository extends JpaRepository<Machine, Integer> {
     @Modifying(clearAutomatically = true)
     @Query(value = "truncate table Machine", nativeQuery = true)
