+++ src/main/java/burgerput/project/zenput/repository/masterAccountRepository/MasterAccountRepository.java
@@ -4,7 +4,6 @@ import burgerput.project.zenput.domain.MasterAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
 public interface MasterAccountRepository extends JpaRepository<MasterAccount, String> {
 
 }
